refactor: remove blazegraph sparql adapter and support selection with relations
