+++ src/index.ts
@@ -1,11 +1,6 @@
 // Mapping
 export * from './mapping/Mapper';
 
-// Storage/Blazegraph
-export * from './storage/blazegraph/BlazegraphQueryAdapter';
-export * from './storage/blazegraph/BlazegraphQueryAdapterOptions';
-export * from './storage/blazegraph/BlazegraphQueryBuilder';
-
 // Storage/Memory
 export * from './storage/memory/MemoryQueryAdapter';
 export * from './storage/memory/MemoryQueryAdapterOptions';
@@ -23,11 +18,9 @@ export * from './storage/operator/LessThanOrEqual';
 export * from './storage/operator/Not';
 
 // Storage/Sparql
-export * from './storage/sparql/BasicSparqlQueryAdapter';
-export * from './storage/sparql/BasicSparqlQueryBuilder';
+export * from './storage/sparql/SparqlQueryAdapter';
 export * from './storage/sparql/SparqlQueryBuilder';
 export * from './storage/sparql/SparqlQueryExecutor';
-export * from './storage/sparql/SparqlQueryPatternBuilder';
 export * from './storage/sparql/SparqlUpdateBuilder';
 export * from './storage/sparql/VariableGenerator';
